backend/requirements.txt
# ============================================================
# CORE DJANGO & REST FRAMEWORK
# ============================================================
Django>=5.0,<6.1
djangorestframework>=3.15.0,<4.0
asgiref>=3.8.0
tzdata>=2024.1

# ============================================================
# DATABASE
# ============================================================
# The modern PostgreSQL driver for Django
psycopg[binary]>=3.1.18

# ============================================================
# AUTHENTICATION & SECURITY
# ============================================================
# Powers the JWT /login/ and /refresh/ endpoints we built
djangorestframework-simplejwt>=5.3.1

# Critical for allowing your Next.js/React frontend to talk to this API
django-cors-headers>=4.3.1

# ============================================================
# ENVIRONMENT & SECRETS
# ============================================================
# Powers the manage.py environment loading we set up
python-dotenv>=1.0.1
# Enterprise standard for reading .env into settings.py
django-environ>=0.11.2

# ============================================================
# API DOCUMENTATION
# ============================================================
# Auto-generates the Swagger/ReDoc pages we added to urls.py
drf-spectacular>=0.27.2

# ============================================================
# PRODUCTION SERVER & STATIC FILES
# ============================================================
# The actual web server that will run Django in production (Render, AWS, etc.)
gunicorn>=21.2.0
# Serves static files (like the Admin portal CSS) efficiently without Nginx
whitenoise>=6.6.0